Update (2): Check for new releases

+++ zphisher.sh
@@ -164,23 +164,31 @@ kill_pid() {
 # Check for a newer release
 check_update(){
 	echo -ne "\n${GREEN}[${WHITE}+${GREEN}]${CYAN} Checking for update : "
+	BASE_DIR="$(realpath "$(dirname "$BASH_SOURCE")")"
 	relase_url='https://api.github.com/repos/htr-tech/zphisher/releases/latest'
 	new_version=$(curl -s "${relase_url}" | grep '"tag_name":' | awk -F\" '{print $4}')
-	tarball_url=https://github.com/htr-tech/zphisher/archive/refs/tags/${new_version}.tar.gz
+	tarball_url="https://github.com/htr-tech/zphisher/archive/refs/tags/${new_version}.tar.gz"
 
 	if [[ $new_version != $__version__ ]]; then
-		echo -ne "${ORANGE} update found\n"${WHITE}
+		echo -ne "${ORANGE}update found\n"${WHITE}
 		sleep 2
 		echo -ne "\n${GREEN}[${WHITE}+${GREEN}]${ORANGE} Downloading Update..."
-		pushd "$BASE_DIR/.." > /dev/null 2>&1
+		pushd "$HOME" > /dev/null 2>&1
 		curl --silent --insecure --fail --retry-connrefused \
 		--retry 3 --retry-delay 2 --location --output ".zphisher.tar.gz" "${tarball_url}"
-		[[ -e ".zphisher.tar.gz" ]] && tar -xf .zphisher.tar.gz -C "$BASE_DIR" --strip-components 1 > /dev/null 2>&1
-		rm -f .zphisher.tar.gz
-		popd > /dev/null 2>&1
-		{ sleep 3; clear; banner_small; }
-		echo -ne "\n${GREEN}[${WHITE}+${GREEN}] Successfully updated! Run zphisher again\n\n"${WHITE}
-		{ reset_color ; exit 1; }
+
+		if [[ -e ".zphisher.tar.gz" ]]; then
+			tar -xf .zphisher.tar.gz -C "$BASE_DIR" --strip-components 1 > /dev/null 2>&1
+			[ $? -ne 0 ] && { echo -e "\n\n${RED}[${WHITE}!${RED}]${RED} Error occured while extracting."; reset_color; exit 1; }
+			rm -f .zphisher.tar.gz
+			popd > /dev/null 2>&1
+			{ sleep 3; clear; banner_small; }
+			echo -ne "\n${GREEN}[${WHITE}+${GREEN}] Successfully updated! Run zphisher again\n\n"${WHITE}
+			{ reset_color ; exit 1; }
+		else
+			echo -e "\n${RED}[${WHITE}!${RED}]${RED} Error occured while downloading."
+			{ reset_color; exit 1; }
+		fi
 	else
 		echo -ne "${GREEN}up to date\n${WHITE}" ; sleep .5
 	fi
